Refetch deployment run logs every second depending on the run status
