+++ client/my-sites/github-deployments/deployment-run-logs/use-code-deployment-run-query.ts
@@ -1,9 +1,39 @@
 import { useQuery, UseQueryOptions } from '@tanstack/react-query';
 import wp from 'calypso/lib/wp';
-import { DeploymentRun } from 'calypso/my-sites/github-deployments/deployments/use-code-deployments-query';
 import { GITHUB_DEPLOYMENTS_QUERY_KEY } from '../constants';
+import { CodeDeploymentData } from '../deployments/use-code-deployments-query';
 
 export const CODE_DEPLOYMENTS_RUNS_QUERY_KEY = 'code-deployments-runs';
+
+export type DeploymentRunStatus =
+	| 'pending'
+	| 'queued'
+	| 'running'
+	| 'success'
+	| 'failed'
+	| 'warnings'
+	| 'building'
+	| 'dispatched'
+	| 'unknown';
+
+export interface DeploymentRun {
+	id: number;
+	code_deployment_id: number;
+	created_on: string;
+	started_on: string;
+	completed_on: string;
+	status: DeploymentRunStatus;
+	failure_code: string;
+	triggered_by_user_id: number;
+	metadata: Metadata;
+	code_deployment?: CodeDeploymentData;
+}
+
+export interface Metadata {
+	commit_message: string;
+	commit_sha: string;
+	job_id: number;
+}
 
 export const useCodeDeploymentsRunsQuery = (
 	siteId: number | null,
